requirements.txt
openai>=1.37.0
PyYAML>=6.0.1
python-docx>=1.1.0
pandas>=2.2.2
jinja2>=3.1.4
requests>=2.32.3
